backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
mysql-connector-python>=8.0.0
email-validator>=1.3.0
passlib[bcrypt]>=1.7.4
python-jose>=3.3.0
